requirements.txt
pyside2
jedi
qtmodern
pyyaml